requirements.txt
aioprocessing==2.0.1
alabaster==0.7.13
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
black==23.3.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.5.7
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
coverage==7.3.0
distlib==0.3.7
dnspython==2.3.0
docutils==0.20.1
email-validator==2.0.0.post2
esbonio==0.16.1
fastapi==0.104.1
filelock==3.12.2
flake8==6.1.0
flake8-black==0.3.6
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
identify==2.5.27
idna==3.4
imagesize==1.4.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
livereload==2.6.3
lsprotocol==2023.0.0b1
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.9.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
pycodestyle==2.11.0
pydantic==2.5.2
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.14.5
pyflakes==3.1.0
pygls==1.0.2
Pygments==2.16.1
pyproject-api==1.5.4
pyspellchecker==0.7.2
pytest==7.4.0
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
starlette==0.27.0
tornado==6.3.2
tox==4.10.0
typeguard==3.0.2
typing_extensions==4.9.0
ujson==5.8.0
urllib3==2.0.3
uvloop==0.17.0
virtualenv==20.24.3
watchfiles==0.21.0
websockets==11.0.3
